修正 ・inputコンポネントに分ける ・/profile/profile-setting ⇒ /profile
