+++ front/app/pages/profile.tsx
@@ -1,23 +1,5 @@
 import React from 'react'
-
-type InputProps = {
-    type: string
-    size?: 'large' | 'small'
-    placeholder: string
-    labletext: string
-}
-
-const Input = ({type, size, placeholder, labletext}: InputProps)=> {
-    return (
-        <div className='my-4'>
-            <div className='text-lg'>{labletext}</div>
-            <input type={type} className={`h-12 ${size=='large'?'w-80':'w-40'} text-lg rounded bg-slate-200`} placeholder={placeholder}></input>
-        </div>
-    )
-}
-
-
-
+import Input from 'components/Input'
 const Profile = () => {
     return(
         <div className='flex flex-col mx-auto my-12 w-80'>
@@ -30,11 +12,11 @@ const Profile = () => {
             </div>
             <div className='my-2'>
                 <div className='text-lg m-auto my-4'>基本情報</div>
-                <Input type='text' size='small' placeholder='郵便番号' labletext='郵便番号'></Input>
-                <Input type='text' size='large' placeholder='住所を入力' labletext='住所'></Input>
-                <Input type='text' size='large' placeholder='番地を入力' labletext='番地'></Input>
-                <Input type='text' size='large' placeholder='建物を入力' labletext='建物名'></Input>
-                <Input type='text' size='large' placeholder='電話番号' labletext='電話番号'></Input>
+                <Input type='text' size='small' placeholder='入力してください' labletext='郵便番号'></Input>
+                <Input type='text' size='large' placeholder='入力してください' labletext='住所'></Input>
+                <Input type='text' size='large' placeholder='入力してください' labletext='番地'></Input>
+                <Input type='text' size='large' placeholder='入力してください' labletext='建物名'></Input>
+                <Input type='text' size='large' placeholder='入力してください' labletext='電話番号'></Input>
             </div>
             <button type='submit' className='my-2 mx-auto w-48 h-12 bg-cyan-400 rounded-full'>決定する</button>
             <button type='submit' className='my-2 mx-auto w-36 h-8 text-stone-400 border-solid border-2 rounded-full'>＜　やめる　</button>
